re-org for tokenizer
