+++ package.json
@@ -16,7 +16,8 @@
         "lint": "eslint . && prettier . --check",
         "lint.fix": "eslint . --fix && prettier . --write",
         "validate-ddg-pixel-defs": "node ./bin/validate_schema.mjs",
-        "validate-ddg-live-pixel": "node ./bin/validate_live_pixel.mjs"
+        "validate-ddg-live-pixel": "node ./bin/validate_live_pixel.mjs",
+        "preprocess-defs": "node ./live_validation_scripts/preprocess_defs.mjs"
     },
     "license": "Apache-2.0",
     "devDependencies": {
