+++ src/constants.mjs
@@ -1,0 +1,1 @@
+export const ROOT_PREFIX = "ROOT_PREFIX";
